+++ src/infixaParaPosfixa.h
@@ -1,0 +1,9 @@
+#ifndef INFIXA_POSFIXA_H
+#define INFIXA_POSFIXA_H
+
+#include "basics.h"
+#include "fila.h"
+
+Fila* infixa_para_posfixa(Fila* fila_infixa);
+
+#endif
