+++ src/interface.h
@@ -1,0 +1,9 @@
+#ifndef INTERFACE_H
+#define INTERFACE_H
+
+#include "basics.h"
+
+void exibir_cabecalho();
+void obter_expressao(char* buffer, int tamanho);
+
+#endif
